+++ tests/test_stemma.py
@@ -3,7 +3,7 @@ Unit tests for the Stemma class.
 """
 import unittest
 import numpy as np
-from stemma.stemma import Stemma
+from stemmabench.algorithms.stemma import Stemma
 
 class TestStemma(unittest.TestCase):
     """Unit tests for the Utils functions.
@@ -12,12 +12,16 @@ class TestStemma(unittest.TestCase):
         """Setup the unit test.
         """
         self.stemma_empty = Stemma()
-
-        
+        self.stemma_edge = Stemma()
+        self.stemma_edge.compute("test_data/test_edge.txt")
 
     def test_stemma_fitted_generation(self):
         """Checks that the fitted value is initialised to false."""
         self.assertFalse(self.stemma_empty.fitted)
+
+    def test_stemma_root_generation(self):
+        """Checks that the root value is initialised to None."""
+        self.assertFalse(self.stemma_empty.root)
 
     def test_stemma_lookup_generation(self):
         """Checks that the lookup dictionary is initialized to an empy dict."""
@@ -34,4 +38,11 @@ class TestStemma(unittest.TestCase):
     def test_stemma_generation_info_generation(self):
         """Checks that the generation_info is initialized to None."""
         self.assertIsNone(self.stemma_empty.generation_info)
+
+    def test_stemma_fitted_compute_edge(self):
+        "Check to see if the compute method sets the fitted attribute properly using edge files."
+        self.assertTrue(self.stemma_edge.fitted)
         
+    def test_stemma_lookup_compute_edge(self):
+        "Check to see if the compute method sets the lookup attribute properly using edge files."
+        self.assertDictEqual(self.stemma_edge.lookup, )
